Move ChunkedData and IndexedData to common package
