+++ packages/common/src/evo/data_converters/common/part_utils.py
