+++ packages/omf/src/evo/data_converters/omf/exporter/__init__.py
@@ -15,7 +15,6 @@ from .evo_lineset_to_omf import export_omf_lineset
 from .evo_pointset_to_omf import export_omf_pointset
 from .evo_surface_to_omf import export_omf_surface
 from .evo_to_omf import UnsupportedObjectError, export_omf
-from .utils import ChunkedData, IndexedData
 
 __all__ = [
     "export_blocksync_omf",
@@ -24,7 +23,5 @@ __all__ = [
     "export_omf_lineset",
     "export_omf_pointset",
     "export_omf_surface",
-    "ChunkedData",
-    "IndexedData",
     "UnsupportedObjectError",
 ]
