+++ packages/omf/src/evo/data_converters/omf/exporter/evo_lineset_to_omf.py
@@ -18,10 +18,10 @@ from evo_schemas.objects import LineSegments_V2_0_0, LineSegments_V2_1_0
 from omf import LineSetElement, LineSetGeometry
 from omf.data import ProjectElementData
 
+from evo.data_converters.common.part_utils import ChunkedData
 from evo.objects.utils.data import ObjectDataClient
 
 from .evo_attributes_to_omf import export_omf_attributes
-from .utils import ChunkedData
 
 
 def export_omf_lineset(
